add course

+++ src/app/courses/courses-list/courses-list.component.ts
@@ -6,7 +6,7 @@ import { CommonModule } from '@angular/common';
 @Component({
   selector: 'app-courses-list',
   standalone: true,
-  imports: [RouterModule],
+  imports: [RouterModule,CommonModule],
   templateUrl: './courses-list.component.html',
   styleUrl: './courses-list.component.css'
 })
@@ -38,6 +38,7 @@ export class CoursesListComponent  implements OnInit {
   onSelectCourse(course: Course): void {
     this.selectedCourse = course;
   }
+  
 
   onAddCourse(newCourse: Course): void {
     this.courseService.addCourse(newCourse).subscribe(
@@ -50,15 +51,22 @@ export class CoursesListComponent  implements OnInit {
     );
   }
 
-  onDeleteCourse(courseId: string): void {
+  onDeleteCourse(courseId: string | undefined): void {
+    if (!courseId) return; // Guard against undefined
+  
     this.courseService.deleteCourse(courseId).subscribe(
       () => {
-        this.courses = this.courses.filter((course) => course.id !== courseId);
+       
+        this.courses = this.courses.filter(course => course.id !== courseId);
+        
+        $('#deleteEmployeeModal').modal('hide');
       },
       (error) => {
         console.error('Error deleting course:', error);
       }
     );
   }
+  
+  
   
 }
